feat(code-editor): added code samples that can be easily copied

+++ modules/code-editor/src/backend/editor.ts
@@ -9,7 +9,7 @@ import { Config } from '../config'
 import { FileDefinition, FileTypes } from './definitions'
 import { EditorError } from './editorError'
 import { EditableFile, FilePermissions, FilesDS, FileType, TypingDefinitions } from './typings'
-import { assertValidFilename, buildRestrictedProcessVars, filterBuiltin, getFileLocation } from './utils'
+import { assertValidFilename, buildRestrictedProcessVars, getBuiltinExclusion, getFileLocation } from './utils'
 
 export const FILENAME_REGEX = /^[0-9a-zA-Z_\-.]+$/
 export const MAIN_GLOBAL_CONFIG_FILES = ['botpress.config.json', 'workspaces.json']
@@ -32,10 +32,13 @@ export default class Editor {
     await Promise.mapSeries(Object.keys(permissions), async type => {
       const userPermissions = permissions[type]
       if (userPermissions.read) {
-        const loadedFiles = await this.loadFiles(userPermissions.type, !userPermissions.isGlobal && this._botId)
-        files[type] = this._config.includeBuiltin ? loadedFiles : filterBuiltin(loadedFiles)
+        files[type] = await this.loadFiles(userPermissions.type, !userPermissions.isGlobal && this._botId)
       }
     })
+
+    const examples = await this._getExamples()
+    files['action_example'] = examples.filter(x => x.type === 'action')
+    files['hook_example'] = examples.filter(x => x.type === 'hook')
 
     return files
   }
@@ -67,20 +70,39 @@ export default class Editor {
       return []
     }
 
+    const excluded = this._config.includeBuiltin ? undefined : getBuiltinExclusion()
     const ghost = botId ? this.bp.ghost.forBot(botId) : this.bp.ghost.forGlobal()
     const files = def.filenames
       ? def.filenames
-      : await ghost.directoryListing(baseDir, def.isJSON ? '*.json' : '*.js', undefined, true)
+      : await ghost.directoryListing(baseDir, def.isJSON ? '*.json' : '*.js', excluded, true)
 
     return Promise.map(files, async (filepath: string) => ({
       name: path.basename(filepath),
       type: fileTypeId as FileType,
       location: filepath,
-      // When not including builtin files, we need to read all of them to filter
-      content: !this._config.includeBuiltin ? await ghost.readFileAsString(baseDir, filepath) : undefined,
+      content: undefined,
       botId,
       ...(dirListingAddFields && dirListingAddFields(filepath))
     }))
+  }
+
+  private async _getExamples(): Promise<EditableFile[]> {
+    const files = await this.bp.ghost.forGlobal().directoryListing('/examples', '*.js')
+
+    return Promise.map(files, async (filepath: string) => {
+      const isHook = filepath.startsWith('examples/hooks')
+      const location = filepath.replace('examples/actions/', '').replace('examples/hooks/', '')
+
+      return {
+        name: path.basename(filepath),
+        type: (isHook ? 'hook' : 'action') as FileType,
+        location,
+        readOnly: true,
+        isExample: true,
+        content: await this.bp.ghost.forGlobal().readFileAsString('/examples', filepath),
+        ...(isHook && { hookType: location.substr(0, location.indexOf('/')) })
+      }
+    })
   }
 
   private _getGhost(file: EditableFile): sdk.ScopedGhostService {
@@ -88,6 +110,11 @@ export default class Editor {
   }
 
   async deleteFile(file: EditableFile): Promise<void> {
+    const fileDef = FileTypes[file.type]
+    if (fileDef.canDelete && !fileDef.canDelete(file)) {
+      throw new Error('This file cannot be deleted.')
+    }
+
     const { folder, filename } = getFileLocation(file)
     await this._getGhost(file).deleteFile(folder, filename)
   }
@@ -121,12 +148,16 @@ export default class Editor {
       'utf-8'
     )
 
+    // Required so array.includes() can be used without displaying an error
+    const es6include = fs.readFileSync(path.join(__dirname, '/../typings/es6include.txt'), 'utf-8')
+
     this._typings = {
       'process.d.ts': buildRestrictedProcessVars(),
       'node.d.ts': nodeTyping.toString(),
       'botpress.d.ts': sdkTyping.toString().replace(`'botpress/sdk'`, `sdk`),
       'bot.config.schema.json': botSchema.toString(),
-      'botpress.config.schema.json': botpressConfigSchema.toString()
+      'botpress.config.schema.json': botpressConfigSchema.toString(),
+      'es6include.d.ts': es6include.toString()
     }
 
     return this._typings
